chore: break out sell asset input from SharedTradeInputBody
