+++ src/components/MultiHopTrade/components/LimitOrder/components/LimitOrderBuyAsset.tsx
@@ -1,0 +1,3 @@
+export const LimitOrderBuyAsset = () => {
+  return <> </>
+}
