+++ src/components/MultiHopTrade/components/SharedTradeInput/SharedTradeInputBody.tsx
@@ -8,46 +8,31 @@ import {
   Stack,
 } from '@chakra-ui/react'
 import type { AccountId } from '@shapeshiftoss/caip'
-import type { TradeQuote } from '@shapeshiftoss/swapper'
 import type { Asset } from '@shapeshiftoss/types'
-import { positiveOrZero } from '@shapeshiftoss/utils'
 import { useCallback, useEffect, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { TradeAssetSelect } from 'components/AssetSelection/AssetSelection'
-import { useInputOutputDifferenceDecimalPercentage } from 'components/MultiHopTrade/hooks/useInputOutputDifference'
 import { useAccountsFetchQuery } from 'context/AppProvider/hooks/useAccountsFetchQuery'
 import { useModal } from 'hooks/useModal/useModal'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { useWalletSupportsChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
 import { isToken } from 'lib/utils'
 import {
-  selectHasUserEnteredAmount,
   selectHighestMarketCapFeeAsset,
   selectIsAccountMetadataLoadingByAccountId,
   selectWalletConnectedChainIds,
 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
-import { TradeAssetInput } from '../TradeAssetInput'
 import { SellAssetInput } from '../TradeInput/components/SellAssetInput'
 
-const formControlProps = {
-  borderRadius: 0,
-  background: 'transparent',
-  borderWidth: 0,
-}
 const arrowDownIcon = <ArrowDownIcon />
-const emptyPercentOptions: number[] = []
 
 type SharedTradeInputBodyProps = {
-  activeQuote: TradeQuote | undefined
-  buyAmountAfterFeesCryptoPrecision: string | undefined
-  buyAmountAfterFeesUserCurrency: string | undefined
   buyAsset: Asset
-  buyAssetAccountId: AccountId | undefined
+  children: JSX.Element
   isInputtingFiatSellAmount: boolean
   isLoading: boolean | undefined
-  manualReceiveAddress: string | undefined
   sellAmountCryptoPrecision: string
   sellAmountUserCurrency: string | undefined
   sellAsset: Asset
@@ -55,21 +40,15 @@ type SharedTradeInputBodyProps = {
   handleSwitchAssets: () => void
   onChangeIsInputtingFiatSellAmount: (isInputtingFiatSellAmount: boolean) => void
   onChangeSellAmountCryptoPrecision: (sellAmountCryptoPrecision: string) => void
-  setBuyAsset: (asset: Asset) => void
-  setBuyAssetAccountId: (accountId: AccountId) => void
   setSellAsset: (asset: Asset) => void
   setSellAssetAccountId: (accountId: AccountId) => void
 }
 
 export const SharedTradeInputBody = ({
-  activeQuote,
-  buyAmountAfterFeesCryptoPrecision,
-  buyAmountAfterFeesUserCurrency,
   buyAsset,
-  buyAssetAccountId,
+  children,
   isInputtingFiatSellAmount,
   isLoading,
-  manualReceiveAddress,
   sellAmountCryptoPrecision,
   sellAmountUserCurrency,
   sellAsset,
@@ -77,8 +56,6 @@ export const SharedTradeInputBody = ({
   handleSwitchAssets,
   onChangeIsInputtingFiatSellAmount,
   onChangeSellAmountCryptoPrecision,
-  setBuyAsset,
-  setBuyAssetAccountId,
   setSellAsset,
   setSellAssetAccountId,
 }: SharedTradeInputBodyProps) => {
@@ -91,14 +68,13 @@ export const SharedTradeInputBody = ({
 
   const walletConnectedChainIds = useAppSelector(selectWalletConnectedChainIds)
   const defaultSellAsset = useAppSelector(selectHighestMarketCapFeeAsset)
-  const hasUserEnteredAmount = useAppSelector(selectHasUserEnteredAmount)
   const { isFetching: isAccountsMetadataLoading } = useAccountsFetchQuery()
   const isAccountMetadataLoadingByAccountId = useAppSelector(
     selectIsAccountMetadataLoadingByAccountId,
   )
 
   const walletSupportsBuyAssetChain = useWalletSupportsChain(buyAsset.chainId, wallet)
-  const buyAssetSearch = useModal('buyTradeAssetSearch')
+
   const sellAssetSearch = useModal('sellTradeAssetSearch')
 
   const percentOptions = useMemo(() => {
@@ -107,9 +83,6 @@ export const SharedTradeInputBody = ({
 
     return [1]
   }, [sellAsset.assetId])
-
-  const inputOutputDifferenceDecimalPercentage =
-    useInputOutputDifferenceDecimalPercentage(activeQuote)
 
   // If the user disconnects the chain for the currently selected sell asset, switch to the default asset
   useEffect(() => {
@@ -138,13 +111,6 @@ export const SharedTradeInputBody = ({
     })
   }, [sellAssetSearch, setSellAsset])
 
-  const handleBuyAssetClick = useCallback(() => {
-    buyAssetSearch.open({
-      onAssetClick: setBuyAsset,
-      title: 'trade.tradeTo',
-    })
-  }, [buyAssetSearch, setBuyAsset])
-
   const sellTradeAssetSelect = useMemo(
     () => (
       <TradeAssetSelect
@@ -155,18 +121,6 @@ export const SharedTradeInputBody = ({
       />
     ),
     [handleSellAssetClick, sellAsset.assetId, setSellAsset],
-  )
-
-  const buyTradeAssetSelect = useMemo(
-    () => (
-      <TradeAssetSelect
-        assetId={buyAsset.assetId}
-        onAssetClick={handleBuyAssetClick}
-        onAssetChange={setBuyAsset}
-        onlyConnectedChains={false}
-      />
-    ),
-    [buyAsset.assetId, handleBuyAssetClick, setBuyAsset],
   )
 
   // disable switching assets if the buy asset isn't supported
@@ -225,29 +179,7 @@ export const SharedTradeInputBody = ({
 
         <Divider />
       </Flex>
-      <TradeAssetInput
-        // Disable account selection when user set a manual receive address
-        isAccountSelectionHidden={Boolean(manualReceiveAddress)}
-        isReadOnly={true}
-        accountId={buyAssetAccountId}
-        assetId={buyAsset.assetId}
-        assetSymbol={buyAsset.symbol}
-        assetIcon={buyAsset.icon}
-        cryptoAmount={
-          hasUserEnteredAmount ? positiveOrZero(buyAmountAfterFeesCryptoPrecision).toFixed() : '0'
-        }
-        fiatAmount={
-          hasUserEnteredAmount ? positiveOrZero(buyAmountAfterFeesUserCurrency).toFixed() : '0'
-        }
-        percentOptions={emptyPercentOptions}
-        showInputSkeleton={isLoading}
-        showFiatSkeleton={isLoading}
-        label={translate('trade.youGet')}
-        onAccountIdChange={setBuyAssetAccountId}
-        formControlProps={formControlProps}
-        labelPostFix={buyTradeAssetSelect}
-        inputOutputDifferenceDecimalPercentage={inputOutputDifferenceDecimalPercentage}
-      />
+      {children}
     </Stack>
   )
 }
